Refactor pooling-configuration react queries to supavisor-configuration
