+++ apps/studio/components/interfaces/Connect/ConnectTabContent.tsx
@@ -5,7 +5,7 @@ import { useParams } from 'common'
 import { GenericSkeletonLoader } from 'components/ui/ShimmeringLoader'
 import { useProjectSettingsV2Query } from 'data/config/project-settings-v2-query'
 import { usePgbouncerConfigQuery } from 'data/database/pgbouncer-config-query'
-import { useSupavisorConfigurationQuery } from 'data/database/pooling-configuration-query'
+import { useSupavisorConfigurationQuery } from 'data/database/supavisor-configuration-query'
 import { pluckObjectFields } from 'lib/helpers'
 import { cn } from 'ui'
 import type { projectKeys } from './Connect.types'
