+++ apps/studio/components/interfaces/Settings/Database/ConnectionPooling/ConnectionPooling.tsx
@@ -17,8 +17,8 @@ import { useMaxConnectionsQuery } from 'data/database/max-connections-query'
 import { usePgbouncerConfigQuery } from 'data/database/pgbouncer-config-query'
 import { usePgbouncerConfigurationUpdateMutation } from 'data/database/pgbouncer-config-update-mutation'
 import { usePgbouncerStatusQuery } from 'data/database/pgbouncer-status-query'
-import { useSupavisorConfigurationQuery } from 'data/database/pooling-configuration-query'
-import { useSupavisorConfigurationUpdateMutation } from 'data/database/pooling-configuration-update-mutation'
+import { useSupavisorConfigurationQuery } from 'data/database/supavisor-configuration-query'
+import { useSupavisorConfigurationUpdateMutation } from 'data/database/supavisor-configuration-update-mutation'
 import { useOrgSubscriptionQuery } from 'data/subscriptions/org-subscription-query'
 import { useProjectAddonsQuery } from 'data/subscriptions/project-addons-query'
 import { useCheckPermissions } from 'hooks/misc/useCheckPermissions'
