--- apps/studio/data/database/pooling-configuration-query.ts
@@ -1,70 +1,0 @@
-import { useQuery, UseQueryOptions } from '@tanstack/react-query'
-
-import type { components } from 'data/api'
-import { get, handleError } from 'data/fetchers'
-import type { ResponseError } from 'types'
-import { databaseKeys } from './keys'
-
-export type PoolingConfigurationVariables = {
-  projectRef?: string
-}
-
-export type PoolingConfiguration = components['schemas']['SupavisorConfigResponse']
-
-/**
- * Should rename this to Supavisor eventually since we're supporting both types of poolers
- */
-export async function getPoolingConfiguration(
-  { projectRef }: PoolingConfigurationVariables,
-  signal?: AbortSignal
-) {
-  if (!projectRef) throw new Error('Project ref is required')
-
-  const { data, error } = await get(`/platform/projects/{ref}/config/supavisor`, {
-    params: { path: { ref: projectRef } },
-    signal,
-  })
-  if (error) handleError(error)
-  return data
-}
-
-export type PoolingConfigurationData = Awaited<ReturnType<typeof getPoolingConfiguration>>
-export type PoolingConfigurationError = ResponseError
-
-/**
- * @deprecated use useSupavisorConfigurationQuery isntead
- */
-export const usePoolingConfigurationQuery = <TData = PoolingConfigurationData>(
-  { projectRef }: PoolingConfigurationVariables,
-  {
-    enabled = true,
-    ...options
-  }: UseQueryOptions<PoolingConfigurationData, PoolingConfigurationError, TData> = {}
-) =>
-  useQuery<PoolingConfigurationData, PoolingConfigurationError, TData>(
-    databaseKeys.poolingConfiguration(projectRef),
-    ({ signal }) => getPoolingConfiguration({ projectRef }, signal),
-    {
-      enabled: enabled && typeof projectRef !== 'undefined',
-      ...options,
-    }
-  )
-
-/**
- * Just a duplicate of usePoolingConfigurationQuery until we move everything over
- */
-export const useSupavisorConfigurationQuery = <TData = PoolingConfigurationData>(
-  { projectRef }: PoolingConfigurationVariables,
-  {
-    enabled = true,
-    ...options
-  }: UseQueryOptions<PoolingConfigurationData, PoolingConfigurationError, TData> = {}
-) =>
-  useQuery<PoolingConfigurationData, PoolingConfigurationError, TData>(
-    databaseKeys.poolingConfiguration(projectRef),
-    ({ signal }) => getPoolingConfiguration({ projectRef }, signal),
-    {
-      enabled: enabled && typeof projectRef !== 'undefined',
-      ...options,
-    }
-  )
